cdk-deployment/lambda/requirements.txt
pillow==10.3.0
boto3==1.34.119